+++ src/main/java/Repositories/RepositoriesInterfaces/DepartementRepositoryInterface.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface DepartementRepositoryInterface {
     void save(Departement departement);
-    Departement findById(Long id);
+    Departement findById(int id);
     List<Departement> findAll();
     void update(Departement departement);
     void delete(int id);
